+++ src/main/java/handling/world/CharacterTransfer.java
@@ -1,19 +1,19 @@
 package handling.world;
 
 import client.*;
-import java.io.Externalizable;
-import java.io.IOException;
-import java.io.ObjectInput;
-import java.io.ObjectOutput;
-import java.util.LinkedHashMap;
-import java.util.Map;
-
 import client.inventory.MapleMount;
 import client.inventory.MaplePet;
 import server.quest.MapleQuest;
 import tools.Pair;
+
+import java.io.Externalizable;
+import java.io.IOException;
+import java.io.ObjectInput;
+import java.io.ObjectOutput;
 import java.util.ArrayList;
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 
 public class CharacterTransfer implements Externalizable {
 
@@ -55,7 +55,7 @@ public class CharacterTransfer implements Externalizable {
         this.vpoints = chr.getVPoints();
         this.name = chr.getName();
         this.fame = chr.getFame();
-        this.gender = (byte) chr.getGender();
+        this.gender = chr.getGender();
         this.level = chr.getLevel();
         this.str = chr.getStat().getStr();
         this.dex = chr.getStat().getDex();
@@ -80,8 +80,8 @@ public class CharacterTransfer implements Externalizable {
         this.marriageId = chr.getMarriageId();
         this.world = chr.getWorld();
         this.guildid = chr.getGuildId();
-        this.guildrank = (byte) chr.getGuildRank();
-        this.alliancerank = (byte) chr.getAllianceRank();
+        this.guildrank = chr.getGuildRank();
+        this.alliancerank = chr.getAllianceRank();
         this.gmLevel = (byte) chr.getGMLevel();
         this.points = chr.getPoints();
         this.fairyExp = chr.getFairyExp();
