+++ src/main/java/server/maps/Event_DojoAgent.java
@@ -3,8 +3,6 @@
 */
 package server.maps;
 
-import java.awt.Point;
-
 import client.MapleCharacter;
 import handling.channel.ChannelServer;
 import handling.world.MaplePartyCharacter;
@@ -13,6 +11,8 @@ import server.Timer.MapTimer;
 import server.custom.bossrank.BossRankManager;
 import server.life.MapleLifeFactory;
 import tools.MaplePacketCreator;
+
+import java.awt.*;
 
 public class Event_DojoAgent {
     //工作人员强化特别训练场
@@ -123,7 +123,7 @@ public class Event_DojoAgent {
         try {
             final MapleMap currentmap = c.getMap();
             final int temp = (currentmap.getId() - 925000000) / 100;
-            final int thisStage = (int) (temp - ((temp / 100) * 100));
+            final int thisStage = temp - ((temp / 100) * 100);
             final int points = getDojoPoints(thisStage);
 
             final ChannelServer ch = c.getClient().getChannelServer();
